support blocker condition based on next location
This change modifies the `when` prop to support a function with `currentLocation` (type Location), `nextLocation` (type Location), and `_action` (type Action) parameters that should return true when the navigation to `nextLocation` should be blocked and false when navigation to `nextLocation` should be allowed.
This means that the `when` prop is now the same as the `when` prop from react-router-navigation-prompt.